# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YAOpenEvidence API",
    "description": "YAOpenEvidence 医学文献证据问答 API。\n\n- 问答任务异步执行：`POST /v1/answers` 返回 202，进度走 `GET /v1/jobs/{id}/events`（SSE）\n- 鉴权：管理员创建账号，登录后使用 `Authorization: Bearer <access_token>`（包括 SSE）\n- 错误：RFC 9457 Problem Details（`application/problem+json`），按 `code` 分支\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/kb/search": {
      "get": {
        "tags": [
          "kb"
        ],
        "summary": "检索知识库",
        "operationId": "search_v1_kb_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "fact",
                    "paragraph"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 8,
              "title": "Top K"
            }
          },
          {
            "name": "pmid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Pmid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KbSearchResult"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/kb/stats": {
      "get": {
        "tags": [
          "kb"
        ],
        "summary": "读取知识库统计",
        "operationId": "stats_v1_kb_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KbStats"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/kb/reindex": {
      "post": {
        "tags": [
          "kb"
        ],
        "summary": "重建知识库索引",
        "operationId": "reindex_v1_kb_reindex_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "ask",
              "kb_reindex"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "created_at"
        ],
        "title": "Job"
      },
      "JobError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "JobError"
      },
      "JobProgress": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "stage"
        ],
        "title": "JobProgress"
      },
      "KbHit": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "fact",
              "paragraph"
            ],
            "title": "Kind"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "year": {
            "type": "string",
            "title": "Year",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "quartile": {
            "type": "string",
            "title": "Quartile",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "default": ""
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "sec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sec"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "text_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Zh"
          },
          "fact_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Kind"
          },
          "quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote"
          },
          "verified": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "text",
          "score"
        ],
        "title": "KbHit"
      },
      "KbSearchResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KbHit"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "query",
          "items"
        ],
        "title": "KbSearchResult"
      },
      "KbStats": {
        "properties": {
          "items": {
            "type": "integer",
            "title": "Items"
          },
          "papers": {
            "type": "integer",
            "title": "Papers"
          },
          "by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Kind"
          },
          "embedder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedder"
          },
          "dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dim"
          }
        },
        "type": "object",
        "required": [
          "items",
          "papers",
          "by_kind"
        ],
        "title": "KbStats"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details。`code` 是稳定枚举，`detail` 仅供人读。",
        "properties": {
          "type": {
            "examples": [
              "urn:yaoe:error:not_found"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "examples": [
              "answer 'abc' not found"
            ],
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "examples": [
              "/v1/answers/abc"
            ],
            "title": "Instance",
            "type": "string"
          },
          "code": {
            "examples": [
              "not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "登录返回的可撤销会话令牌",
        "scheme": "bearer"
      }
    }
  }
}